# docs/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: mandarine.app@yandex.ru
    name: Mandarine Support
  description: API for web and mobile application Mandarine
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Mandarine API
  version: 0.0.0
paths:
...
  /v0/masters/profiles/-/services:
    get:
      description: Request for finding master services. User must be logged in. In
        response will be returned found master services.
      operationId: FindMasterServices
      parameters:
        - in: query
          name: field
          schema:
            type: string
        - in: query
          name: maxInterval
          schema:
            type: string
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: minInterval
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Found master services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.MasterServicesOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Find master services
      tags:
        - Master Service API
...
  "/v0/masters/profiles/{username}/services":
    get:
      description: Request for finding master services by username. User must be
        logged in. In response will be returned found master services.
      operationId: FindMasterServicesByUsername
      parameters:
        - description: Username
          in: path
          name: username
          required: true
          schema:
            type: string
        - in: query
          name: field
          schema:
            type: string
        - in: query
          name: maxInterval
          schema:
            type: string
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: minInterval
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Found master services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.MasterServicesOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted; Master profile is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Master profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Find master services by username
      tags:
        - Master Service API
    post:
      description: Request for creating master service. User must be logged in. In
        response will be returned created master service.
      operationId: CreateMasterService
      parameters:
        - description: Username
          in: path
          name: username
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.CreateMasterServiceInput"
        description: Create master service request body
        required: true
      responses:
        "201":
          description: Created master service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.MasterServiceOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted; Master profile is disabled; Cannot
            create master service for another master profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Master service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Create master service
      tags:
        - Master Service API
  "/v0/masters/profiles/{username}/services/{id}":
    delete:
      description: Request for deleting master service. User must be logged in.
      operationId: DeleteMasterService
      parameters:
        - description: Username
          in: path
          name: username
          required: true
          schema:
            type: string
        - description: Master service ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted; Master profile is disabled; Cannot
            delete master service for another master profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Master service not found; Master profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Delete master service
      tags:
        - Master Service API
    get:
      description: Request for getting master service. User must be logged in. In
        response will be returned found master service.
      operationId: GetMasterServiceByUsername
      parameters:
        - description: Username
          in: path
          name: username
          required: true
          schema:
            type: string
        - description: Master profile ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Found master service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.MasterServiceOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted; Master profile is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Master service not found; Master profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Get master service
      tags:
        - Master Service API
    patch:
      description: Request for updating master service. User must be logged in. In
        response will be returned updated master service.
      operationId: UpdateMasterService
      parameters:
        - description: Username
          in: path
          name: username
          required: true
          schema:
            type: string
        - description: Master service ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v0.UpdateMasterServiceInput"
        description: Update master service request body
        required: true
      responses:
        "200":
          description: Updated master service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.MasterServiceOutput"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "403":
          description: User is blocked or deleted; Master profile is disabled; Cannot
            update master service for another master profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "404":
          description: Master service not found; Master profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v0.ErrorOutput"
      security:
        - BearerAuth: []
      summary: Update master service
      tags:
        - Master Service API
...
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
  schemas:
...
    v0.CreateMasterServiceInput:
      properties:
        avatarId:
          type: string
        description:
          type: string
        maxInterval:
          format: hh:mm:ss
          type: string
        maxPrice:
          type: number
        minInterval:
          format: hh:mm:ss
          type: string
        minPrice:
          type: number
        name:
          type: string
      required:
        - name
      type: object
    v0.ErrorOutput:
      properties:
        message:
          type: string
        path:
          example: /api/v0/example
          format: url_path
          type: string
        status:
          maximum: 599
          minimum: 400
          type: integer
        timestamp:
          type: string
      required:
        - message
        - path
        - status
        - timestamp
      type: object
...
    v0.MasterServiceOutput:
      properties:
        avatarId:
          type: string
        description:
          type: string
        id:
          type: string
        maxInterval:
          type: string
        maxPrice:
          type: number
        minInterval:
          type: string
        minPrice:
          type: number
        name:
          type: string
      type: object
    v0.MasterServicesOutput:
      properties:
        count:
          type: integer
        data:
          items:
            $ref: "#/components/schemas/v0.MasterServiceOutput"
          type: array
      type: object
...
    v0.UpdateMasterServiceInput:
      properties:
        avatarId:
          type: string
        description:
          type: string
        maxInterval:
          format: hh:mm:ss
          type: string
        maxPrice:
          type: number
        minInterval:
          format: hh:mm:ss
          type: string
        minPrice:
          type: number
        name:
          type: string
      required:
        - name
      type: object
...